Fix URL building bug
Any call to the buildZapRequestUrl function with 'params' causes the function to return a URL with an extra ampersand ('&') appended to the end. This leads to http error 400 - Bad Request Error. 

The reason this happens is that the ampersand is appended to every map entry in the params Map, including the last one. 

This PR fixes this issue. It is based on the accepted answer to this question
http://stackoverflow.com/questions/3395286/remove-last-character-of-a-stringbuilder 
and is an elegant solution.