Use name for alerts
Change classes Alert and AlertTask to use name instead of alert for the
name of the alert, deprecate the alert setters/getters and add new ones
for the name.
Change other classes to use the new methods.

Related to zaproxy/zaproxy#1341 - Update alert XML schema to remove
element-level name collision