Allow to specify the API key through the ClientApi
Change ClientApi to allow to specify the ZAP API key when creating the
instance, to ensure that the API key is sent with all API requests.
Also, send it in the header not just as (query) parameter (for newer and
older versions of ZAP).
Update SimpleExample to indicate the new usage.

Related to zaproxy/zaproxy#3226 - Option to supply API key or nonces via
header