Change value type of ApiResponseSet to ApiResponse
Change ApiResponseSet to not assume that the values of the set are of
the type ApiResponseElement (by changing the type of the values to
ApiResponse) and add helper method to return the values as String.
Change Alert to use the new method (to get the value(s) as String).
Change ApiResponseElement to return its value when toString() is called.

Fix zaproxy/zaproxy#3228 - ApiResponseElement casting error in
constructor of ApiResponseSet (if child is also a set)