Do not require to set the API key in each call
Change API implementations to not require to set/use the API key in each
API call (e.g. action/other) as that's already done by the ClientApi
class. The old methods (the ones that have the apiKey parameter) are now
in a separate class, to avoid the generated code to replace them, they
are still required for compatibility with older versions. The usage of
the old methods is discouraged (deprecated) and suggested to use the new
ClientApi constructors to set the API key.
Remove usage of the old methods throughout the code (ClientApi,
ClientApiMain, examples, and Ant tasks).