Update Ant tasks to wait for the scans to finish
Change ActiveScanSubtreeTask, ActiveScanUrlTask, and SpiderUrlTask to
wait for the scans to finish before returning.
Refactor the active scan tasks to reduce code duplication (they differ
only on the parameters used to start the scan).
Update the tests to assert the new behaviour (i.e. scan tasks check the
status of the scans).