Check binary compatibility as part of check task
Change japicmp task to run as part of the check task, set it to have as
input file the jar (to ensure it runs, always, when jar is changed), add
a task description and add it to verification tasks. Also, update the
plugin to latest version available.
Set BC version to latest released version.