Update Gradle and move to Maven Publish Plugin
Update Gradle to 4.8.1, change to use Maven Publish Plugin (instead of
Maven Plugin, which is superseded), and remove unnecessary task
generatePom, the publish plugin already provides a task to generate it.
Update OWASP Dependency-Check plugin to 3.2.1 and Spotless to 3.13.0.
Update RELEASING.md with the new task to publish the artifacts.